requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
colorama==0.4.6
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
environs==11.0.0
executing==2.0.1
inflection==0.5.1
ipython==8.25.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
marshmallow==3.21.2
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pillow==10.3.0
prompt_toolkit==3.0.46
psycopg==3.1.19
pure-eval==0.2.2
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
wcwidth==0.2.13
fpdf2
